chore: write tests

+++ src/task.hook.ts
@@ -1,17 +1,17 @@
 import fs from "fs";
-import path from "path";
 import { PNG } from "pngjs";
 import pixelmatch, { PixelmatchOptions } from "pixelmatch";
 import moveFile from "move-file";
-import sanitize from "sanitize-filename";
+import path from "path";
+import { FILE_SUFFIX, IMAGE_SNAPSHOT_PREFIX, TASK } from "./constants";
 import {
-  FILE_SUFFIX,
-  IMAGE_SNAPSHOT_PREFIX,
-  TASK,
-  PATH_VARIABLES,
-  WINDOWS_LIKE_DRIVE_REGEX,
-} from "./constants";
-import { alignImagesToSameSize, importAndScaleImage } from "./image.utils";
+  cleanupUnused,
+  alignImagesToSameSize,
+  scaleImageAndWrite,
+  isImageCurrentVersion,
+  writePNG,
+} from "./image.utils";
+import { generateScreenshotPath } from "./screenshotPath.utils";
 import type { CompareImagesTaskReturn } from "./types";
 
 export type CompareImagesCfg = {
@@ -31,34 +31,22 @@ const unlinkSyncSafe = (path: string) =>
 const moveSyncSafe = (pathFrom: string, pathTo: string) =>
   fs.existsSync(pathFrom) && moveFile.sync(pathFrom, pathTo);
 
-export const getScreenshotPathTask = ({
-  title,
-  imagesPath,
-  specPath,
-}: {
-  title: string;
+export const getScreenshotPathInfoTask = (cfg: {
+  titleFromOptions: string;
   imagesPath: string;
   specPath: string;
 }) => {
-  const parsePathPartVariables = (pathPart: string, i: number) => {
-    if (pathPart === PATH_VARIABLES.specPath) {
-      return path.dirname(specPath);
-    } else if (i === 0 && !pathPart) {
-      // when unix-like absolute path
-      return PATH_VARIABLES.unixSystemRootPath;
-    } else if (i === 0 && WINDOWS_LIKE_DRIVE_REGEX.test(pathPart)) {
-      // when win-like absolute path
-      return path.join(PATH_VARIABLES.winSystemRootPath, pathPart[0]);
-    }
+  const screenshotPath = generateScreenshotPath(cfg);
 
-    return pathPart;
-  };
+  return { screenshotPath, title: path.basename(screenshotPath, ".png") };
+};
 
-  return path.join(
-    IMAGE_SNAPSHOT_PREFIX,
-    ...imagesPath.split("/").map(parsePathPartVariables),
-    `${sanitize(title)}${FILE_SUFFIX.actual}.png`
-  );
+export const cleanupImagesTask = (config: Cypress.PluginConfigOptions) => {
+  if (config.env["pluginVisualRegressionCleanupUnusedImages"]) {
+    cleanupUnused(config.projectRoot);
+  }
+
+  return null;
 };
 
 export const approveImageTask = ({ img }: { img: string }) => {
@@ -77,16 +65,18 @@ export const compareImagesTask = async (
   cfg: CompareImagesCfg
 ): Promise<CompareImagesTaskReturn> => {
   const messages = [] as string[];
+  const rawImgNewBuffer = await scaleImageAndWrite({
+    scaleFactor: cfg.scaleFactor,
+    path: cfg.imgNew,
+  });
   let imgDiff: number | undefined;
   let imgNewBase64: string, imgOldBase64: string, imgDiffBase64: string;
   let error = false;
 
   if (fs.existsSync(cfg.imgOld) && !cfg.updateImages) {
-    const rawImgNew = await importAndScaleImage({
-      scaleFactor: cfg.scaleFactor,
-      path: cfg.imgNew,
-    });
-    const rawImgOld = PNG.sync.read(fs.readFileSync(cfg.imgOld));
+    const rawImgNew = PNG.sync.read(rawImgNewBuffer);
+    const rawImgOldBuffer = fs.readFileSync(cfg.imgOld);
+    const rawImgOld = PNG.sync.read(rawImgOldBuffer);
     const isImgSizeDifferent =
       rawImgNew.height !== rawImgOld.height ||
       rawImgNew.width !== rawImgOld.width;
@@ -130,7 +120,7 @@ export const compareImagesTask = async (
     imgOldBase64 = PNG.sync.write(imgOld).toString("base64");
 
     if (error) {
-      fs.writeFileSync(
+      writePNG(
         cfg.imgNew.replace(FILE_SUFFIX.actual, FILE_SUFFIX.diff),
         diffBuffer
       );
@@ -144,8 +134,13 @@ export const compareImagesTask = async (
         maxDiffThreshold: cfg.maxDiffThreshold,
       };
     } else {
-      // don't overwrite file if it's the same (imgDiff < cfg.maxDiffThreshold && !isImgSizeDifferent)
-      fs.unlinkSync(cfg.imgNew);
+      if (rawImgOld && !isImageCurrentVersion(rawImgOldBuffer)) {
+        writePNG(cfg.imgNew, rawImgNewBuffer);
+        moveFile.sync(cfg.imgNew, cfg.imgOld);
+      } else {
+        // don't overwrite file if it's the same (imgDiff < cfg.maxDiffThreshold && !isImgSizeDifferent)
+        fs.unlinkSync(cfg.imgNew);
+      }
     }
   } else {
     // there is no "old screenshot" or screenshots should be immediately updated
@@ -153,6 +148,7 @@ export const compareImagesTask = async (
     imgNewBase64 = "";
     imgDiffBase64 = "";
     imgOldBase64 = "";
+    writePNG(cfg.imgNew, rawImgNewBuffer);
     moveFile.sync(cfg.imgNew, cfg.imgOld);
   }
 
@@ -182,8 +178,9 @@ export const doesFileExistTask = ({ path }: { path: string }) =>
   fs.existsSync(path);
 
 /* c8 ignore start */
-export const initTaskHook = () => ({
-  [TASK.getScreenshotPath]: getScreenshotPathTask,
+export const initTaskHook = (config: Cypress.PluginConfigOptions) => ({
+  [TASK.getScreenshotPathInfo]: getScreenshotPathInfoTask,
+  [TASK.cleanupImages]: cleanupImagesTask.bind(undefined, config),
   [TASK.doesFileExist]: doesFileExistTask,
   [TASK.approveImage]: approveImageTask,
   [TASK.compareImages]: compareImagesTask,
